# ls: Cache sort keys for default sort order